ActionFormWrapper component
